requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2